fixed relative links
I made a mistake with how relative links work. I removed the path_prefix from the link href to avoid redirecting to non-existant directories. I'm keeping the path_prefix variable in the code for now because I intend to expose it in the config file along with some other stuff like dates