inheritance fix
moved config loading out to the recursive step. originally everything was done in one big function and we were able to exploit scope for inheritance. But when we moved recursion into its own step we left config loading inside the build step meaning that recursive iterations did not have the loaded config as a parent scope.

also added a test string to the test-directory. since the inheritance issue was missed because the test originally used the default config as root